addressed pr comments - restructured theme and made minor changes in SearchBar and IconTextField
