+++ client/src/App.jsx
@@ -1,10 +1,9 @@
 import React, { Suspense } from "react";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
-import { createTheme, CssBaseline, ThemeProvider } from "@mui/material";
+import { CssBaseline, ThemeProvider } from "@mui/material";
 import "@fontsource/poppins";
 
-import text from "./theme/text";
-import color from "./theme/color";
+import theme from "./theme";
 
 const withSuspense = (Component) => (
   <Suspense fallback="brewing...">
@@ -13,24 +12,6 @@ const withSuspense = (Component) => (
 );
 
 const HomePage = React.lazy(() => import("./pages/HomePage"));
-
-const theme = createTheme({
-  palette: {
-    ...color,
-  },
-  typography: {
-    ...text,
-  },
-  components: {
-    MuiTypography: {
-      styleOverrides: {
-        root: {
-          fontFamily: "Poppins",
-        },
-      },
-    },
-  },
-});
 
 const App = () => {
   return (
