--- client/src/theme/color.js
@@ -1,52 +1,0 @@
-// TODO: Add brand colors map
-// eslint-disable-next-line
-const BRAND_COLORS = {
-  BLUEBERRY: "#021944",
-  LYCHEE: "#F6A794",
-  MATCHA: "#B1D5B2",
-  MILK_TEA: "#F2DAC4",
-  TARO: "#E4CFF7",
-  THAI_TEA: "#E0A878",
-};
-
-const color = {
-  brown: {
-    50: "#FAF0E7",
-    100: BRAND_COLORS.MILK_TEA,
-    200: "#E0A878",
-    300: "#A4857B",
-    400: "#98513A",
-    500: "#683323",
-  },
-  gray: {
-    50: "#FFFFFF",
-    100: "#F7EDEA",
-    200: "#E4E2E0",
-    300: "#8C8C8C",
-    400: "#E7D8CE",
-    500: "#D9D9D9",
-    600: "#B4AAA0",
-    700: "#979797",
-    800: "#2F3032",
-    900: "#000000",
-  },
-  red: {
-    50: "#F8D7D7",
-    100: "#F4D0BD",
-    200: "#E2BCA9",
-    300: "#F6A794",
-    400: "#D21C1C",
-    500: "#B03737",
-  },
-  blue: {
-    50: "#ABBFE3",
-    100: "#005EA2",
-    200: "#021944",
-  },
-  green: {
-    50: "#73956F",
-    100: "#489879",
-  },
-};
-
-export default color;
