--- client/src/theme/text.js
@@ -1,16 +1,0 @@
-const text = {
-  color: "#2F3032",
-  fontFamily: "Poppins",
-  h1: { fontWeight: "bold", fontSize: "4.5rem" },
-  h2: { fontWeight: "bold", fontSize: "3.5rem" },
-  h3: { fontWeight: "bold", fontSize: "3rem" },
-  h4: { fontWeight: "bold", fontSize: "2rem" },
-  h5: { fontWeight: "bold", fontSize: "1.5rem" },
-  h6: { fontWeight: "bold", fontSize: "1rem" },
-  title: { fontWeight: "600", fontSize: "3.5rem", lineHeight: "5.5rem" },
-  subtitle: { fontSize: "1.125rem" },
-  body1: { fontSize: "1.125rem" },
-  body2: { fontSize: "1.125rem", color: "#C4C4C4" },
-};
-
-export default text;
